requirements.txt
Pillow
pdf2image
PyInstaller